Added unittest and some changes
