+++ setup.cfg
@@ -1,2 +1,32 @@
+[flake8]
+select = C, E, W, F, I
+max-line-length = 88
+extend-ignore = E203, W503
+max-complexity = 10
+exclude = __init__.py
+
+[coverage:run]
+source =
+    src
+    docs/src
+    tests
+branch = True
+parallel = True
+
+[coverage:report]
+precision = 2
+fail_under = 98.09
+skip_covered = True
+show_missing = True
+exclude_lines =
+    @abc.abstractmethod
+omit =
+    docs/src/logging/tutorial003.py
+    docs/src/logging/tutorial004.py
+    docs/src/logging/tutorial005.py
+
+[isort]
+profile = black
+
 [metadata]
 description-file = README.md
