+++ test/test_clearance_iscwsa.py
@@ -1,7 +1,10 @@
-from welleng.survey import Survey, make_survey_header
-from welleng.clearance import Clearance, ISCWSA
-import numpy as np
 import json
+import numpy as np
+import os
+import unittest
+
+from welleng.clearance import Clearance, ISCWSA
+from welleng.survey import Survey, make_survey_header
 
 """
 Test that the ISCWSA clearance model is working within a defined tolerance
@@ -13,14 +16,46 @@ error model.
 # Set test tolerance as percentage
 TOLERANCE = 0.5
 
-# Read well and validation data
-filename = (
-    "tests/test_data/clearance_iscwsa_well_data.json"
-)
-data = json.load(open(filename))
+
+class TestClearanceISCWSA(unittest.TestCase):
+
+    def test_clearance_iscwsa(self):
+
+        # Read well and validation data
+        filename = os.path.abspath(
+            os.path.join(
+                os.path.dirname(__file__),
+                'test_data', 'clearance_iscwsa_well_data.json'
+            )
+        )
+        data = json.load(open(filename))
+
+        surveys = generate_surveys(data)
+        reference = surveys["Reference well"]
+
+        # Perform clearance checks for each survey
+        for well in surveys:
+            if well == "Reference well":
+                continue
+            else:
+                offset = surveys[well]
+                # skip well 10
+                if well == "10 - well":
+                    continue
+                else:
+                    c = Clearance(reference, offset)
+
+                result = ISCWSA(c)
+
+                normalized = np.absolute(
+                    np.array(result.SF) - np.array(data["wells"][well]["SF"])
+                ) / np.array(data["wells"][well]["SF"]) * 100
+
+                assert np.all(normalized < TOLERANCE)
 
 
 def generate_surveys(data):
+
     # Generate surveys for imported wells
     surveys = {}
 
@@ -58,33 +93,3 @@ def generate_surveys(data):
         surveys[well] = s
 
     return surveys
-
-
-def test_clearance_iscwsa(data=data, tolerance=TOLERANCE):
-    surveys = generate_surveys(data)
-    reference = surveys["Reference well"]
-
-    # Perform clearance checks for each survey
-    for well in surveys:
-        if well == "Reference well":
-            continue
-        else:
-            offset = surveys[well]
-            # skip well 10
-            if well == "10 - well":
-                continue
-            else:
-                c = Clearance(reference, offset)
-
-            result = ISCWSA(c)
-
-            normalized = np.absolute(
-                np.array(result.SF) - np.array(data["wells"][well]["SF"])
-            ) / np.array(data["wells"][well]["SF"]) * 100
-
-            assert np.all(normalized < tolerance)
-
-
-# make above test runnanble separately
-if __name__ == '__main__':
-    test_clearance_iscwsa(data=data, tolerance=TOLERANCE)
