+++ test/test_data/extract_clearance_model_well_data.py
