+++ test/test_data/extract_iscwsa_error_data.py
