+++ welleng/survey.py
@@ -21,10 +21,11 @@ from .utils import (
     get_xyz,
     radius_from_dls
 )
-from .error import ErrorModel, ERROR_MODELS
-from .node import Node
-from .connector import Connector, interpolate_well
-from .visual import figure
+
+from welleng.error import ErrorModel, ERROR_MODELS
+from welleng.node import Node
+from welleng.connector import Connector, interpolate_well
+from welleng.visual import figure
 
 
 AZI_REF = ["true", "magnetic", "grid"]
@@ -49,11 +50,7 @@ class SurveyHeader:
         deg=True,
         depth_unit='meters',
         surface_unit='meters',
-        mag_defaults={
-            'b_total': 50_000.,
-            'dip': 70.,
-            'declination': 0.,
-        },
+        mag_defaults=None,
         vertical_section_azimuth=0,
         **kwargs
     ):
@@ -147,7 +144,11 @@ class SurveyHeader:
         self.azi_reference = azi_reference
         self.vertical_section_azimuth = vertical_section_azimuth
 
-        self.mag_defaults = mag_defaults
+        self.mag_defaults = mag_defaults or {
+            'b_total': 50_000.,
+            'dip': 70.,
+            'declination': 0.,
+        }
         self._get_mag_data(deg)
 
     def _get_mag_data(self, deg):
@@ -242,8 +243,8 @@ class Survey:
         cov_nev=None,
         cov_hla=None,
         error_model=None,
-        start_xyz=[0., 0., 0.],
-        start_nev=[0., 0., 0.],
+        start_xyz=None,
+        start_nev=None,
         start_cov_nev=None,
         deg=True,
         unit="meters",
@@ -338,8 +339,8 @@ class Survey:
 
         self.unit = unit
         self.deg = deg
-        self.start_xyz = start_xyz
-        self.start_nev = start_nev
+        self.start_xyz = start_xyz if start_xyz is not None else np.array([0., 0., 0.])
+        self.start_nev = start_nev if start_nev is not None else np.array([0., 0., 0.])
         self.md = np.array(md).astype('float64')
         self.start_cov_nev = start_cov_nev
 
@@ -399,7 +400,9 @@ class Survey:
                 self.azi_true_deg - math.degrees(self.header.declination)
             )
             self._get_azi_mag_and_true_rad()
+
         elif self.header.azi_reference == 'true':
+
             if deg:
                 self.azi_true_deg = np.array(azi).astype('float64')
                 self.azi_mag_deg = (
@@ -407,6 +410,7 @@ class Survey:
                 )
                 self._get_azi_mag_and_true_rad()
                 azi_temp = self._get_azi_temp(deg)
+
             else:
                 self.azi_true_rad = np.array(azi).astype('float64')
                 self.azi_mag_rad = (
@@ -415,7 +419,9 @@ class Survey:
                 self._get_azi_mag_and_true_deg()
                 azi_temp = self._get_azi_temp(deg)
             self._make_angles(inc, azi_temp, deg)
+
         else:  # azi_reference is "magnetic"
+
             if deg:
                 self.azi_mag_deg = np.array(azi).astype('float64')
                 self.azi_true_deg = (
@@ -423,6 +429,7 @@ class Survey:
                 )
                 self._get_azi_mag_and_true_rad()
                 azi_temp = self._get_azi_temp(deg)
+
             else:
                 self.azi_mag_rad = np.array(azi).astype('float64')
                 self.azi_true_rad = (
@@ -527,8 +534,8 @@ class Survey:
 
         if return_error:
             return self.err
-        else:
-            return self
+
+        return self
 
     def _get_errors(self):
         """
@@ -543,11 +550,14 @@ class Survey:
             )
             self.cov_hla = self.err.errors.cov_HLAs.T
             self.cov_nev = self.err.errors.cov_NEVs.T
+
         else:
             if self.cov_nev is not None and self.cov_hla is None:
                 self.cov_hla = NEV_to_HLA(self.survey_rad, self.cov_nev.T).T
+
             elif self.cov_nev is None and self.cov_hla is not None:
                 self.cov_nev = HLA_to_NEV(self.survey_rad, self.cov_hla.T).T
+
             else:
                 pass
 
@@ -562,10 +572,7 @@ class Survey:
         with np.errstate(divide='ignore', invalid='ignore'):
             radius = 1 / curvature
         circumference = 2 * np.pi * radius
-        if self.unit == 'meters':
-            x = 30
-        else:
-            x = 100
+        x = 30 if self.unit == 'meters' else 100
         rate = np.absolute(np.degrees(2 * np.pi / circumference) * x)
 
         return rate
@@ -578,10 +585,7 @@ class Survey:
         # split the survey
         s = SplitSurvey(self)
 
-        if self.unit == 'meters':
-            x = 30
-        else:
-            x = 100
+        x = 30 if self.unit == 'meters' else 100
 
         # this is lazy I know, but I'm using this mostly for flags
         with np.errstate(divide='ignore', invalid='ignore'):
@@ -669,7 +673,7 @@ class Survey:
 
         Examples
         --------
-        >>> import welleng as we
+        >>> import src as we
         >>> survey = we.connector.interpolate_survey(
         ...    survey=we.survey.Survey(
         ...       md=[0, 500, 1000, 2000, 3000],
@@ -747,6 +751,7 @@ class Survey:
         """
         if dls is None:
             dls = self.dls[-1]
+
         if toolface is None:
             toolface = self.toolface[-1]
 
@@ -911,10 +916,10 @@ def interpolate_md(survey, md):
     x = md - survey.md[idx]
     assert x >= 0
 
-    return _interpolate_survey(survey, x=x, index=idx)
+    return _interpolate_survey(survey, delta_md=x, index=idx)
 
 
-def _interpolate_survey(survey, x=0, index=0):
+def _interpolate_survey(survey, delta_md=0, index=0):
     """
     Interpolates a point distance x between two survey stations
     using minimum curvature.
@@ -923,7 +928,7 @@ def _interpolate_survey(survey, x=0, index=0):
     ----------
         survey: welleng.Survey
             A survey object with at least two survey stations.
-        x: float
+        delta_md: float
             Length along well path from indexed survey station to
             perform the interpolate at. Must be less than length
             to the next survey station.
@@ -955,7 +960,7 @@ def _interpolate_survey(survey, x=0, index=0):
 
         total_dogleg = survey.dogleg[index + 1]
 
-        dogleg = x * (total_dogleg / survey.delta_md[index + 1])
+        dogleg = delta_md * (total_dogleg / survey.delta_md[index + 1])
 
         t = (
             (math.sin(total_dogleg - dogleg) / math.sin(total_dogleg)) * t1
@@ -970,7 +975,7 @@ def _interpolate_survey(survey, x=0, index=0):
     sh.azi_reference = 'grid'
 
     s = Survey(
-        md=np.array([survey.md[index], survey.md[index] + x]),
+        md=np.array([survey.md[index], survey.md[index] + delta_md]),
         inc=np.array([survey.inc_rad[index], inc]),
         azi=np.array([survey.azi_grid_rad[index], azi]),
         start_xyz=np.array([survey.x, survey.y, survey.z]).T[index],
@@ -979,8 +984,8 @@ def _interpolate_survey(survey, x=0, index=0):
         deg=False,
     )
     interpolated = False if any((
-        x == 0,
-        x == survey.md[index + 1] - survey.md[index]
+        delta_md == 0,
+        delta_md == survey.md[index + 1] - survey.md[index]
      )) else True
     s.interpolated = [False, interpolated]
 
@@ -1026,7 +1031,7 @@ def interpolate_tvd(survey, tvd, **kwargs):
         dogleg = s_temp.dogleg[-1]
 
     if np.isnan(dogleg):
-        return _interpolate_survey(survey, x=0, index=idx)
+        return _interpolate_survey(survey, delta_md=0, index=idx)
 
     if dogleg == 0:
         x = (
@@ -1083,7 +1088,7 @@ def interpolate_tvd(survey, tvd, **kwargs):
 
         assert x <= delta_md
 
-    interpolated_survey = _interpolate_survey(survey, x=x, index=idx)
+    interpolated_survey = _interpolate_survey(survey, delta_md=x, index=idx)
 
     interpolated = True if x > 0 else False
     node = get_node(interpolated_survey, 1, interpolated=interpolated)
